refactor out example new (annotated) data dict fixture into conftest.py
